renamed and moved example files
