+++ tsconfig.json
@@ -54,6 +54,6 @@
         // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
         // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
     },
-    "include": ["src"],
+    "include": ["src", "examples/runFasterCode.ts", "examples/runOldCode.ts"],
     "exclude": ["node_modules", "**/__tests__/*"]
 }
